+++ src/core_codemods/semgrep/api.py
@@ -1,12 +1,6 @@
-from functools import cache
-from pathlib import Path
-
 from codemodder.codemods.base_codemod import Metadata, Reference, ToolMetadata, ToolRule
-from codemodder.codemods.base_detector import BaseDetector
 from codemodder.codemods.base_transformer import BaseTransformerPipeline
-from codemodder.context import CodemodExecutionContext
-from codemodder.result import ResultSet
-from codemodder.semgrep import SemgrepResultSet
+from codemodder.codemods.semgrep import SemgrepSarifFileDetector
 from core_codemods.api.core_codemod import CoreCodemod, SASTCodemod
 
 
@@ -49,25 +43,3 @@ class SemgrepCodemod(SASTCodemod):
             detector=SemgrepSarifFileDetector(),
             requested_rules=[rule_id],
         )
-
-
-class SemgrepSarifFileDetector(BaseDetector):
-    def apply(
-        self,
-        codemod_id: str,
-        context: CodemodExecutionContext,
-        files_to_analyze: list[Path],
-    ) -> ResultSet:
-        del codemod_id
-        del files_to_analyze
-        return process_semgrep_findings(
-            tuple(context.tool_result_files_map.get("semgrep", ()))
-        )  # Convert list to tuple for cache hashability
-
-
-@cache
-def process_semgrep_findings(semgrep_sarif_files: tuple[str]) -> ResultSet:
-    results = SemgrepResultSet()
-    for file in semgrep_sarif_files or ():
-        results |= SemgrepResultSet.from_sarif(file)
-    return results
