+++ tests/codemods/semgrep/test_semgrep_enable_jinja2_autoescape.py
@@ -11,7 +11,7 @@ class TestEnableJinja2Autoescape(BaseSASTCodemodTest):
     tool = "semgrep"
 
     def test_name(self):
-        assert self.codemod.name == "enable-jinja2-autoescape-semgrep"
+        assert self.codemod.name == "enable-jinja2-autoescape"
 
     def test_import(self, tmpdir):
         input_code = """
